+++ flecsi/execution.hh
@@ -162,8 +162,10 @@ struct trace {
   /// \deprecated Use the default constructor.
   explicit trace(id_t id);
 
-  /// Default move constructor.
-  trace(trace &&) = default;
+  /// Traces are movable.  Those that have been moved from must not be used.
+  trace(trace &&) noexcept;
+  /// Traces can be (move-)assigned.
+  trace & operator=(trace) & noexcept;
 
   struct guard;